fix: modifiers on recently edited files
