+++ event-log/src/main/scala/io/renku/eventlog/subscriptions/package.scala
@@ -18,7 +18,6 @@
 
 package io.renku.eventlog
 
-import cats.Semigroup
 import ch.datascience.graph.model.projects
 import ch.datascience.tinytypes.constraints.Url
 import ch.datascience.tinytypes.json.TinyTypeDecoders.stringDecoder
@@ -35,20 +34,5 @@ package object subscriptions {
 
   trait SubscriptionCategoryPayload {
     def subscriberUrl: SubscriberUrl
-  }
-
-  sealed trait SubscriptionResult
-  final case object SuccesfulSubscription extends SubscriptionResult
-  final case class UnsupportedPayload(message: String) extends SubscriptionResult
-  final case object NoCategoriesAvailable extends SubscriptionResult
-
-  sealed trait RegistrationResult
-  final case object AcceptedRegistration extends RegistrationResult
-  final case object RejectedRegistration extends RegistrationResult
-
-  implicit val registrationResultSemigroup: Semigroup[RegistrationResult] = {
-    case (RejectedRegistration, RejectedRegistration) => RejectedRegistration
-    case _                                            => AcceptedRegistration
-
   }
 }
