+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/SubscriptionCategoryRegistrySpec.scala
@@ -23,6 +23,8 @@ import cats.syntax.all._
 import ch.datascience.generators.Generators.Implicits._
 import ch.datascience.generators.Generators._
 import io.circe.Json
+import io.renku.eventlog.subscriptions.SubscriptionCategory._
+import io.renku.eventlog.subscriptions.SubscriptionCategoryRegistry._
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
